Notifications refactoring
